Improve graphics_triangle_fill
I noticed that large filled triangles (100s of pixels on a side) really slows down the frame rate of the graphics.    100s of small triangles does not seem to be a burden, but 1 or 2 large triangles (say to make a rectangle to fill half the screen) bring the frame rate from 15-20Hz down to 2-3Hz.  Using the filled rectangle function of an equivalent size has no ill effect on the refresh speed.  I’ve attached a file that has a triangle fill algorithm I wrote that appears to be about twice as fast as the current triangle fill in graphics.h.  I suggest a replacement.  graphics_triangle_f1is the current algorithm taken from github.  f2 is my proposed replacement function.